Opendaq fix structures in variable types (#4)
* created with Variable Type BaseDataVariableType.

If creating a structure with BaseDataVariableType as type definition,
the attribute read before has no value. Because of this writeAttribute()
is not called.

If creating a structure with another variable type as type definition,
the attribute has a value. Strange it tells, that it is an extension
object and not a structure anymore.
Afterwards writeAttribute() in the NodeSet changes the structre to an extension object.

This leads to the problem that we can not describe and handle the node properly after construction.

Here we fix the problem by not overwriting the attribute if it of type extension object.

* fix indentation